Reorganise code into packages.
