+++ src/main/java/duke/commands/Command.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.commands;
+
+import duke.Parser;
 
 import java.util.function.Consumer;
 
@@ -12,7 +14,7 @@ public class Command {
 
     /**
      * Executes the command. This runs the lambda previously passed to it in the constructor.
-     * @param args the Parser object that contains the args passed to the command in the user's input. It is assumed 
+     * @param args the Parser object that contains the args passed to the command in the user's input. It is assumed
      *             that the parser has already read the first token (i.e. the name of the command).
      */
     public void run(Parser args) {
@@ -20,10 +22,10 @@ public class Command {
     }
 
     /**
-     * Command constructor method. 
+     * Command constructor method.
      * @param name The name of the command. It should be the name that is used by the user to run that command while
      *             The program is running.
-     * @param executor A lambda that takes a <code>Parser</code> object. The executor should be written with the 
+     * @param executor A lambda that takes a <code>Parser</code> object. The executor should be written with the
      *                 assumption that the parser has already read one token.
      */
     public Command(String name, Consumer<Parser> executor) {
