+++ src/main/java/duke/tasks/TaskList.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.tasks;
 
 import java.util.ArrayList;
 import java.util.stream.Stream;
