+++ src/main/java/duke/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.tasks;
 
 public class ToDo extends Task {
     public ToDo(String name) {
