+++ src/main/java/duke/ui/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.ui;
 
 import java.util.Scanner;
 
@@ -6,9 +6,9 @@ public class Ui {
 
     /** Scanner for user input. */
     private final Scanner input = new Scanner(System.in);
-    
+
     /**
-     * Prints a message to the terminal, decorated with the Louie icon. 
+     * Prints a message to the terminal, decorated with the Louie icon.
      * Printing an empty string just outputs the icon
      *
      * @param message The message to print.
@@ -30,7 +30,7 @@ public class Ui {
         sb.append("\n");
         System.out.print(sb);
     }
-    
+
     public String readInput() {
         return input.nextLine();
     }
